refactor: move command definitions to src/commands and reduce main.ts to wiring
Extract the 27-entry buildCommands table out of src/main.ts into
src/commands/definitions.ts. Introduce a PluginContext interface in
src/commands/plugin-context.ts that bundles the dependencies
buildCommands needs (kill ring, mark, yank-pop session, logger,
killCtx, commandInvoked). EmacsTextEditorPlugin now declares
'implements PluginContext' so the type is structurally enforced.

This breaks the import cycle that would otherwise arise (definitions.ts
importing main.ts when main.ts already imports definitions.ts).

CommandDef.editorCallback drops its third 'plugin' parameter: the
context is now closed over inside buildCommands, so individual
callbacks just receive (editor, view). registerCommands forwards
accordingly.

src/main.ts shrinks from ~410 lines to 62. It now contains only the
plugin class with its state fields, onload/onunload, the mousedown
hook, and commandInvoked.

Verified: typecheck clean, lint clean, build clean, 29 tests still
passing.

Completes Task 1.8 (extract editor-ops modules).
